Added empty directory checking to cron.sh

+++ cron.sh
@@ -6,20 +6,28 @@
 #####################################
 
 # Configuration
-SCAN_DIR=/tmp/scans/*   # Directory where scans are stored
+SCAN_DIR=/tmp/scans   # Directory where scans are stored
 # End Configuration
 
-# Loop over all of the scans
+# Let the user know we're starting
 echo "Checking for scans older than 30 minutes..."
 
-for file in $SCAN_DIR
-do
-        # Is the scan older than 30 minutes?
-        if [ `stat --format=%Y $file` -le $(( `date +%s` - 1800 )) ]; then
-                # Yes, delete it
-                rm -f $file
-                echo "Deleted: $file"
-        fi
-done
+# Are there any scans to check?
+if [ "$(ls -A $SCAN_DIR)" ]; then
+     # Yes, loop over all of the scans
+	for file in $SCAN_DIR
+	do
+			# Is the scan older than 30 minutes?
+			if [ `stat --format=%Y $file` -le $(( `date +%s` - 1800 )) ]; then
+					# Yes, delete it
+					rm -f $file
+					echo "Deleted: $file"
+			fi
+	done
 
-echo "Done!"
+		# Let the user know we're done processing
+	echo "Done!"
+else
+	# Let the user know there's nothing to do
+    echo "No scans to process."
+fi
